+++ assets/javascripts/kitten/components/cards/reward-card/components/row.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import classNames from 'classnames'
+
+export const RewardCardRow = props => (
+  <div
+    {...props}
+    className={classNames('k-RewardCard__cardRow', props.className)}
+  />
+)