+++ assets/javascripts/kitten/components/cards/reward-card/legacy/content.js
@@ -1,17 +1,15 @@
 import React from 'react'
-import Radium from 'radium'
 import PropTypes from 'prop-types'
-import { styles } from '../../../components/cards/reward-card'
-import { Marger } from '../../../components/layout/marger'
-import { Title } from '../../../components/typography/title'
-import { Button } from '../../../components/buttons/button'
-import { HorizontalStroke } from '../../../components/layout/horizontal-stroke'
-import { Paragraph } from '../../../components/typography/paragraph'
-import { Text } from '../../../components/typography/text'
-import { StarIcon } from '../../../components/icons/star-icon'
-import { Deprecated } from '../../../helpers/utils/deprecated'
+import { Marger } from '../../../../components/layout/marger'
+import { Title } from '../../../../components/typography/title'
+import { Button } from '../../../../components/buttons/button'
+import { HorizontalStroke } from '../../../../components/layout/horizontal-stroke'
+import { Paragraph } from '../../../../components/typography/paragraph'
+import { Text } from '../../../../components/typography/text'
+import { StarIcon } from '../../../../components/icons/star-icon'
+import { Deprecated } from '../../../../helpers/utils/deprecated'
 
-const RewardCardContentBase = ({
+export const RewardCardContent = ({
   titleAmount,
   titleTag,
   subtitle,
@@ -22,11 +20,11 @@ const RewardCardContentBase = ({
   isDisabled,
   isTinyVersion,
 }) => {
-  const descriptionStyle = [isDisabled && styles.disabled]
-
   return (
     <Deprecated warningMessage="Please use RewardCard sub-component to make your composition. You can check some examples on https://kisskissbankbank.github.io/../../../">
-      <div style={descriptionStyle} disabled={isDisabled}>
+      <div
+        className={isDisabled ? 'k-LegacyRewardCard__element--disabled' : null}
+      >
         {starred && (
           <Marger bottom="2">
             <Button
@@ -48,7 +46,7 @@ const RewardCardContentBase = ({
             italic
             margin={false}
             tag={titleTag}
-            style={styles.textColor}
+            className="k-u-color-font1"
           >
             {titleAmount}
           </Title>
@@ -63,7 +61,7 @@ const RewardCardContentBase = ({
               size={isTinyVersion ? 'big' : 'huge'}
               tag={subtitleTag}
               weight="bold"
-              style={styles.textMargin}
+              className="k-u-margin-none"
             >
               {subtitle}
             </Text>
@@ -71,7 +69,7 @@ const RewardCardContentBase = ({
         )}
         <Marger top={!subtitle ? 3 : 1}>
           <Paragraph
-            style={styles.textColor}
+            className="k-u-color-font1"
             modifier={isTinyVersion ? 'quaternary' : 'tertiary'}
             margin={false}
           >
@@ -83,7 +81,7 @@ const RewardCardContentBase = ({
   )
 }
 
-RewardCardContentBase.propTypes = {
+RewardCardContent.propTypes = {
   titleAmount: PropTypes.string.isRequired,
   titleTag: PropTypes.string,
   subtitle: PropTypes.string,
@@ -95,7 +93,7 @@ RewardCardContentBase.propTypes = {
   isTinyVersion: PropTypes.bool.isRequired,
 }
 
-RewardCardContentBase.defaultProps = {
+RewardCardContent.defaultProps = {
   titleTag: 'h1',
   subtitle: '',
   subtitleTag: 'p',
@@ -104,5 +102,3 @@ RewardCardContentBase.defaultProps = {
   starLabel: '',
   isDisabled: false,
 }
-
-export const RewardCardContent = Radium(RewardCardContentBase)