+++ assets/javascripts/kitten/components/cards/reward-card/legacy/infos.js
@@ -1,24 +1,22 @@
 import React from 'react'
-import Radium from 'radium'
 import PropTypes from 'prop-types'
-import { styles } from '../../../components/cards/reward-card'
-import { Info } from '../../../components/cards/reward-card/info'
-import { Marger } from '../../../components/layout/marger'
-import { Deprecated } from '../../../helpers/utils/deprecated'
+import { Info } from './info'
+import { Marger } from '../../../../components/layout/marger'
+import { Deprecated } from '../../../../helpers/utils/deprecated'
 
-const RewardCardInfosBase = ({
+export const RewardCardInfos = ({
   infos,
   isDisabled,
   viewportIsTabletOrLess,
   isTinyVersion,
 }) => {
-  const styleInfos = [isDisabled && styles.disabled]
-
   if (infos.length === 0) return null
 
   return (
     <Deprecated warningMessage="Please use RewardCard sub-component to make your composition. You can check some examples on https://kisskissbankbank.github.io/../../../">
-      <div style={styleInfos} disabled={isDisabled}>
+      <div
+        className={isDisabled ? 'k-LegacyRewardCard__element--disabled' : null}
+      >
         <Marger top="2" bottom="3">
           {infos &&
             infos.map(info => (
@@ -36,16 +34,14 @@ const RewardCardInfosBase = ({
   )
 }
 
-RewardCardInfosBase.propTypes = {
+RewardCardInfos.propTypes = {
   infos: PropTypes.array,
   isDisabled: PropTypes.bool,
   viewportIsTabletOrLess: PropTypes.bool.isRequired,
   isTinyVersion: PropTypes.bool.isRequired,
 }
 
-RewardCardInfosBase.defaultProps = {
+RewardCardInfos.defaultProps = {
   infos: [],
   isDisabled: false,
 }
-
-export const RewardCardInfos = Radium(RewardCardInfosBase)